+++ graphein/ml/conversion.py
@@ -36,8 +36,27 @@ except ImportError:
         conda_channel="dglteam",
     )
 
+try:
+    import jax.numpy as jnp
+except ImportError:
+    import_message(
+        submodule="graphein.ml.conversion",
+        package="jax",
+        pip_install=True,
+        conda_channel="conda-forge",
+    )
+try:
+    import jraph
+except ImportError:
+    import_message(
+        submodule="graphein.ml.conversion",
+        package="jraph",
+        pip_install=True,
+        conda_channel="conda-forge",
+    )
 
-SUPPORTED_FORMATS = ["nx", "pyg", "dgl"]
+
+SUPPORTED_FORMATS = ["nx", "pyg", "dgl", "jraph"]
 """Supported conversion formats.
 
 ``"nx"``: NetworkX graph
@@ -45,6 +64,8 @@ SUPPORTED_FORMATS = ["nx", "pyg", "dgl"]
 ``"pyg"``: PyTorch Geometric Data object
 
 ``"dgl"``: DGL graph
+
+``"Jraph"``: Jraph GraphsTuple
 """
 
 SUPPORTED_VERBOSITY = ["gnn", "default", "all_info"]
@@ -58,10 +79,10 @@ class GraphFormatConvertor:
     formats can be retrieved from :const:`~graphein.ml.conversion.SUPPORTED_FORMATS`.
 
     :param src_format: The type of graph you'd like to convert from. Supported formats are available in :const:`~graphein.ml.conversion.SUPPORTED_FORMATS`
-    :type src_format: Literal["nx", "pyg", "dgl"]
+    :type src_format: Literal["nx", "pyg", "dgl", "jraph"]
     :param dst_format: The type of graph format you'd like to convert to. Supported formats are available in:
         ``graphein.ml.conversion.SUPPORTED_FORMATS``
-    :type dst_format:  Literal["nx", "pyg", "dgl"]
+    :type dst_format:  Literal["nx", "pyg", "dgl", "jraph"]
     :param verbose: Select from ``"gnn"``, ``"default"``, ``"all_info"`` to determine how much information is preserved (features)
         as some are unsupported by various downstream frameworks
     :type verbose: graphein.ml.conversion.SUPPORTED_VERBOSITY
@@ -164,7 +185,7 @@ class GraphFormatConvertor:
         :rtype: dgl.DGLGraph
         """
         g = dgl.DGLGraph()
-        node_id = [n for n in G.nodes()]
+        node_id = list(G.nodes())
         G = nx.convert_node_labels_to_integers(G)
 
         ## add node level feat
@@ -190,11 +211,8 @@ class GraphFormatConvertor:
                 ).type("torch.FloatTensor")
             elif self.type2form[i] == "str":
                 string_dict[i] = j
-            elif self.type2form[i] == "float":
+            elif self.type2form[i] in ["float", "int"]:
                 node_dict_transformed[i] = torch.Tensor(np.array(j))
-            elif self.type2form[i] == "int":
-                node_dict_transformed[i] = torch.Tensor(np.array(j))
-
         g.add_nodes(
             len(node_id),
             node_dict_transformed,
@@ -217,18 +235,16 @@ class GraphFormatConvertor:
         for i, j in node_dict.items():
             if self.type2form[i] == "str":
                 string_dict[i] = j
-            elif self.type2form[i] == "float":
+            elif self.type2form[i] in ["float", "int"]:
                 edge_transform_dict[i] = torch.Tensor(np.array(j))
-            elif self.type2form[i] == "int":
-                edge_transform_dict[i] = torch.Tensor(np.array(j))
-
         g.add_edges(edge_index[0], edge_index[1], edge_transform_dict)
 
         # add graph level features
-        graph_dict = {}
-        for i, feat_name in enumerate(G.graph):
-            if str(feat_name) in self.columns:
-                graph_dict[str(feat_name)] = [G.graph[feat_name]]
+        graph_dict = {
+            str(feat_name): [G.graph[feat_name]]
+            for feat_name in G.graph
+            if str(feat_name) in self.columns
+        }
 
         return g
 
@@ -313,6 +329,61 @@ class GraphFormatConvertor:
         :rtype: nx.Graph
         """
         return torch_geometric.utils.to_networkx(G)
+
+    def convert_nx_to_jraph(self, G: nx.Graph) -> jraph.GraphsTuple:
+        """Converts NetworkX graph (``nx.Graph``) to Jraph GraphsTuple graph. Requires ``jax`` and ``Jraph``.
+
+        :param G: Networkx graph to convert.
+        :type G: nx.Graph
+        :return: Jraph GraphsTuple graph.
+        :rtype: jraph.GraphsTuple
+        """
+        G = nx.convert_node_labels_to_integers(G)
+
+        n_node = len(G)
+        n_edge = G.number_of_edges()
+        edge_list = list(G.edges())
+        senders, receivers = zip(*edge_list)
+        senders, receivers = jnp.array(senders), jnp.array(receivers)
+
+        # Add node features
+        node_features = {}
+        for i, (_, feat_dict) in enumerate(G.nodes(data=True)):
+            for key, value in feat_dict.items():
+                if str(key) in self.columns:
+                    node_features[str(key)] = (
+                        [value]
+                        if i == 0
+                        else node_features[str(key)] + [value]
+                    )
+
+        # Add edge features
+        edge_features = {}
+        for i, (_, _, feat_dict) in enumerate(G.edges(data=True)):
+            for key, value in feat_dict.items():
+                if str(key) in self.columns:
+                    edge_features[str(key)] = (
+                        list(value)
+                        if i == 0
+                        else edge_features[str(key)] + list(value)
+                    )
+
+        # Add graph features
+        global_context = {
+            str(feat_name): [G.graph[feat_name]]
+            for feat_name in G.graph
+            if str(feat_name) in self.columns
+        }
+
+        return jraph.GraphsTuple(
+            nodes=node_features,
+            senders=senders,
+            receivers=receivers,
+            edges=edge_features,
+            n_node=n_node,
+            n_edge=n_edge,
+            globals=global_context,
+        )
 
     def __call__(self, G: nx.Graph):
         nx_g = eval("self.convert_" + self.src_format + "_to_nx(G)")